agent: Split OS-specific build for binary analysis
